🚚 Move VersionList component to /documents/components
